# backend/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "minLength": 24,
        "maxLength": 24,
        "pattern": "^\\d{24}$"
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "googleId": {
            "type": "string"
          },
          "googlePic": {
            "type": "string",
            "format": "uri"
          },
          "googleName": {
            "type": "string"
          },
          "facebookId": {
            "type": "string"
          },
          "facebookPic": {
            "type": "string",
            "format": "uri"
          },
          "facebookName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "info": {
    "title": "MangaBookmark API",
    "version": "3.0.0",
    "description": "API for MangaBookmark client"
  },
  "paths": {
...
    "/api/user/{authProvider}/auth": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "authProvider",
            "description": "specifies which third party to use",
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "facebook"
              ]
            }
          },
          {
            "in": "query",
            "name": "action",
            "description": "specifies which action to take",
            "schema": {
              "type": "string",
              "enum": [
                "register",
                "login",
                "link"
              ]
            }
          }
        ],
        "description": "Authentication using Google/Facebook OAuth2",
        "responses": {
          "302": {
            "description": "Redirect to Google/Facebook authentication page"
          }
        }
      }
    },
    "/api/user/{authProvider}/callback": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "authProvider",
            "description": "specifies which third party to use",
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "facebook"
              ]
            }
          }
        ],
        "description": "Google/Facebook login/register/link callback.",
        "responses": {
          "302": {
            "description": "Success - Redirect to quick access page. Failure - Redirect to login page"
          }
        }
      }
    },
    "/api/user/{authProvider}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "authProvider",
            "description": "specifies which third party to use",
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "facebook"
              ]
            }
          }
        ],
        "description": "Remove Google/Facebook from current account",
        "responses": {
          "200": {
            "description": "Account unlinked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/user": {
      "post": {
        "description": "Create new user with email & password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Get current logged in user info",
        "responses": {
          "200": {
            "description": "Retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "Edit current user info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User info updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete current user",
        "responses": {
          "204": {
            "description": "User deleted successfully"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Log in to the application",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "incorrect username/password"
          }
        }
      }
    },
...
    "/api/user/change-password": {
      "patch": {
        "description": "Change current user password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed successfully"
          }
        }
      }
    }
  },
...
}